requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
node==0.9.25
odict==1.7.0
Pillow==8.2.0
plumber==1.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
zope.component==5.0.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.4.0
zope.lifecycleevent==4.3
